requirements.txt
pyglet
pyglm